requirements.txt
streamlit
pandas
numpy
scikit-learn
requests
transformers
torch
joblib
python-dotenv
